Move all constants to new files.
