+++ assets/src/edit-story/elements/image/index.js
@@ -17,9 +17,6 @@
 /**
  * Internal dependencies
  */
-import PanelTypes from '../../components/panels/design/types';
-import { SHARED_DEFAULT_ATTRIBUTES } from '../shared/constants';
-import { MEDIA_DEFAULT_ATTRIBUTES, MEDIA_PANELS } from '../media/constants';
 export { default as Edit } from './edit';
 export { default as Frame } from './frame';
 export { default as Output } from './output';
@@ -33,17 +30,7 @@ export {
   isMedia,
   hasEditMode,
   editModeGrayout,
+  resizeRules,
 } from '../media';
 
-export { resizeRules } from '../media/constants';
-
-export const defaultAttributes = {
-  ...SHARED_DEFAULT_ATTRIBUTES,
-  ...MEDIA_DEFAULT_ATTRIBUTES,
-};
-
-export const panels = [
-  PanelTypes.ELEMENT_ALIGNMENT,
-  ...MEDIA_PANELS,
-  PanelTypes.IMAGE_ACCESSIBILITY,
-];
+export * from './constants';
