+++ assets/src/edit-story/elements/media/index.js
@@ -48,12 +48,4 @@ export const CropBox = styled.div`
   }
 `;
 
-export const hasEditMode = true;
-
-export const isMedia = true;
-
-export const canFlip = true;
-
-export const isMaskable = true;
-
-export const editModeGrayout = true;
+export * from './constants';
