+++ assets/src/edit-story/elements/shape/constants.js
@@ -16,27 +16,22 @@
 /**
  * Internal dependencies
  */
-import { DEFAULT_ATTRIBUTES_FOR_MEDIA } from '../../constants';
+import { SHARED_DEFAULT_ATTRIBUTES } from '../shared/constants';
+import createSolidFromString from '../../utils/createSolidFromString';
 import PanelTypes from '../../components/panels/design/types';
 
-export const MEDIA_DEFAULT_ATTRIBUTES = {
-  ...DEFAULT_ATTRIBUTES_FOR_MEDIA,
-  resource: {
-    alt: '',
-  },
+export const defaultAttributes = {
+  ...SHARED_DEFAULT_ATTRIBUTES,
+  backgroundColor: createSolidFromString('#ffffff'),
 };
 
-export const MEDIA_MASK_OPACITY = 0.4;
+export const hasEditMode = false;
 
-export const hasEditMode = true;
-
-export const isMedia = true;
+export const isMedia = false;
 
 export const canFlip = true;
 
 export const isMaskable = true;
-
-export const editModeGrayout = true;
 
 export const resizeRules = {
   vertical: true,
@@ -46,12 +41,14 @@ export const resizeRules = {
   minHeight: 20,
 };
 
-export const MEDIA_PANELS = [
-  PanelTypes.BACKGROUND_SIZE_POSITION,
-  PanelTypes.LAYER_STYLE,
+export const panels = [
+  PanelTypes.COLOR_PRESETS,
+  PanelTypes.ELEMENT_ALIGNMENT,
   PanelTypes.SIZE_POSITION,
-  PanelTypes.BORDER_RADIUS,
   PanelTypes.BORDER,
-  PanelTypes.ANIMATION,
+  PanelTypes.LAYER_STYLE,
   PanelTypes.LINK,
+  PanelTypes.SHAPE_STYLE,
+  PanelTypes.BORDER_RADIUS,
+  PanelTypes.ANIMATION,
 ];
