+++ assets/src/edit-story/elements/sticker/index.js
@@ -17,36 +17,9 @@
 /**
  * Internal dependencies
  */
-import PanelTypes from '../../components/panels/design/types';
-import { SHARED_DEFAULT_ATTRIBUTES } from '../shared/constants';
-
 export { default as LayerContent } from './layer';
 export { default as Output } from './output';
 export { default as LayerIcon } from './icon';
 export { default as Display } from './display';
 
-export const hasEditMode = false;
-
-export const isMedia = false;
-
-export const canFlip = true;
-
-export const isMaskable = false;
-
-export const isAspectAlwaysLocked = true;
-
-export const resizeRules = {
-  diagonal: true,
-  minWidth: 20,
-  minHeight: 20,
-};
-
-export const defaultAttributes = SHARED_DEFAULT_ATTRIBUTES;
-
-export const panels = [
-  PanelTypes.ELEMENT_ALIGNMENT,
-  PanelTypes.LAYER_STYLE,
-  PanelTypes.SIZE_POSITION,
-  PanelTypes.ANIMATION,
-  PanelTypes.LINK,
-];
+export * from './constants';
